requirements.txt
# --- Core ---
alpaca-py==0.23.0
polygon-api-client==1.10.0
ta==0.11.0
python-dotenv==1.0.0
websockets==11.0.3         # 12.x needs Py 3.11+

# --- Logging / utilities ---
structlog==24.1.0

# --- Dev / CI ---
pytest==8.2.0              # tests/test_bot.py smoke test
ruff==0.4.4                # lint (optional but fast)

# --- Pin transitive deps that occasionally break builds ---
pytz==2024.1
numpy==1.26.4
pandas==2.2.2
